% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subtypes.R
\name{extract_subtypes}
\alias{extract_subtypes}
\title{Extract subtypes from card's type line}
\usage{
extract_subtypes(type_line, split = FALSE)
}
\arguments{
\item{type_line}{card's type line.}

\item{split}{If the card has multiple subtypes, should
\code{extract_subtypes} return a list where each subtype is its own element
of the vector  or one element with all subtypes
(list(c("Goblin", "Warrior")) vs list("Goblin Warrior")))?}
}
\value{
list containing a character vector with the cards' subtypes
}
\description{
Take a card's type type line (stored in the \code{type_line})
column in the output from \code{\link{scry_cards}}) and get just the subtypes
(everything after the '-')
}
\examples{
extract_subtypes("Creature - Merfolk Wizard")
extract_subtypes("Creature - Merfolk Wizard", split = TRUE)

}
\concept{manipulate}
