% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_guild.R
\name{label_guild}
\alias{label_guild}
\title{Assign guild identity based on color/color ID}
\usage{
label_guild(color_code, inclusive = FALSE, mana_cost = NULL)
}
\arguments{
\item{inclusive}{}
}
\description{
Assign guild identity based on color/color ID
}
\examples{
label_guild(c("U", "W"))
label_guild("U")
label_guild("U", inclusive = TRUE)
}
\concept{label}
