% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_mtg_color.R
\name{relabel_mtg_color}
\alias{relabel_mtg_color}
\title{Turn Scryfall color codes into human-readable labels}
\usage{
relabel_mtg_color(color_code)
}
\arguments{
\item{color_code}{character vector (which can be of length 0) containing
single letter WUBRG color codes.}
}
\value{
For mono-colored cards, \code{relabel_mtg_color} returns the card's
color. Otherwise, it returns "Colorless" or "Multicolored", as appropriate.
}
\description{
Color values from \code{\link{scry_cards}} come in as single
character strings (e.g. "U" for blue cards). \code{relabel_mtg_color}
converts them into human-readable text,  (e.g. "Red" or "Multicolored").
}
\section{Warning}{

\code{relabel_mtg_color} treats any character vector of length >1 as a
multicolored card.
}

\examples{
relabel_mtg_color("G")
relabel_mtg_color(c("G", "U"))

}
\seealso{
\code{\link{wubrg_order}} for converting the output into an ordered
factor.
}
\concept{label}
