% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-catalogs.R
\name{scry-catalogs}
\alias{scry-catalogs}
\alias{scry_catalogs_card_names}
\alias{scry_catalogs_artist_names}
\alias{scry_catalogs_word_bank}
\alias{scry_catalogs_creature_types}
\alias{scry_catalogs_planeswalker_types}
\alias{scry_catalogs_land_types}
\alias{scry_catalogs_artifact_types}
\alias{scry_catalogs_enchantment_types}
\alias{scry_catalogs_spell_types}
\alias{scry_catalogs_powers}
\alias{scry_catalogs_toughnesses}
\alias{scry_catalogs_loyalties}
\alias{scry_catalogs_watermarks}
\alias{scry_catalogs_keyword_abilities}
\alias{scry_catalogs_keyword_actions}
\alias{scry_catalogs_ability_words}
\title{Retrieve catalog objects}
\usage{
scry_catalogs_card_names()

scry_catalogs_artist_names()

scry_catalogs_word_bank()

scry_catalogs_creature_types()

scry_catalogs_planeswalker_types()

scry_catalogs_land_types()

scry_catalogs_artifact_types()

scry_catalogs_enchantment_types()

scry_catalogs_spell_types()

scry_catalogs_powers()

scry_catalogs_toughnesses()

scry_catalogs_loyalties()

scry_catalogs_watermarks()

scry_catalogs_keyword_abilities()

scry_catalogs_keyword_actions()

scry_catalogs_ability_words()
}
\value{
A string vector.
}
\description{
A Catalog object contains an array of Magic datapoints (words, card values,
etc). Catalog objects are provided by the API as aids for building other
Magic software and understanding possible values for a field on Card objects.
}
\references{
\url{https://scryfall.com/docs/api/catalogs}
}
