% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rulings.R
\name{scry-rulings}
\alias{scry-rulings}
\alias{scry_ruling}
\alias{scry_ruling_number}
\title{Retrieve information about rulings}
\usage{
scry_ruling(id, source = c("scryfall", "multiverse", "mtgo", "arena"))

scry_ruling_number(number, set)
}
\arguments{
\item{id}{Unique card indentifier. May refer to distinct \code{source}s.}

\item{source}{Source to which \code{id} refers. Can be any one of \code{scryfall},
\code{multiverse}, \code{mtgo}, or \code{arena}.}

\item{number}{The collector number.}

\item{set}{The three to five-letter set code.}
}
\value{
A data frame with 1 or more rows and the following columns:
\itemize{
\item \code{object} [chr]: A content type for this object, always \code{ruling}.
\item \code{oracle_id} [chr]: Unique identifier on the Oracle database.
\item \code{source} [chr]: A computer-readable string indicating which company
produced this ruling, either wotc or scryfall.
\item \code{published_at} [date]: The date when the ruling or note was published.
\item \code{comment} [tbl]: The text of the ruling.
}
}
\description{
Retrieve information about rulings
}
\details{
Rulings represent Oracle rulings, Wizards of the Coast set release
notes, or Scryfall notes for a particular card.
Rulings with a \code{scryfall} source have been added by the Scryfall team, either
to provide additional context for the card, or explain how the card works in
an unofficial format (such as Duel Commander).
}
