% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-sets.R
\name{scry-sets}
\alias{scry-sets}
\alias{scry_set}
\title{Retrieve information about a set}
\usage{
scry_set(code)
}
\arguments{
\item{code}{The three to five-letter set code. Can be either the \code{code} or
the \code{mtgo_code} for the set.}

\item{id}{The \code{tcgplayer_id} or \code{groupId}.}
}
\value{
A data frame with 1 row and the following columns:
\itemize{
\item \code{object} [String]: Set.
\item \code{id} [UUID]: A unique ID for this set on Scryfall that will not change.
\item \code{code} [String]: The unique three to five-letter code for this set.
\item \code{mtgo_code} [String]: The unique code for this set on MTGO, which may
differ from the regular code.
\item \code{arena_code} [String]: The unique code for this set on MTG Arena, which
may differ from the regular code.
\item \code{tcgplayer_id} [Integer]: This set’s ID on TCGplayer’s API, also known as
the \code{groupId}.
\item \code{name} [String]: The English name of the set.
\item \code{set_type} [String]: A computer-readable classification for this set. See
below.
\item \code{uri} [URI]: A link to this set object on Scryfall’s API.
\item \code{scryfall_uri} [URI]: A link to this set’s permapage on Scryfall’s
\item \code{search_uri} [URI]: A Scryfall API URI that you can request to begin
paginating over the cards in this set.
\item \code{released_at} [Date]: The date the set was released or the first card was
printed in the set (in GMT-8 Pacific time).
\item \code{set_type} [String]: Scryfall provides an overall categorization for each
Set in the \code{set_type} property. An exhaustive list of \code{set_types} is listed
in a separate section below.
\item \code{card_count} [Integer]: The number of cards in this set.
\item \code{parent_set_code} [String]: The set code for the parent set, if any.
\code{promo} and \code{token} sets often have a parent set.
\item \code{printed_size} [Integer]: The denominator for the set’s printed collector
numbers.
\item \code{digital} [Boolean]: True if this set was only released in a video game.
\item \code{nonfoil_only} [Boolean]: True if this set contains only nonfoil cards.
website.
\item \code{foil_only} [Boolean]: True if this set contains only foil cards.
\item \code{block_code} [String]: Nullable  The block code for this set, if any.
\item \code{block} [String]: The block or group name code for this set, if any.
\item \code{icon_svg_uri} [URI]: A URI to an SVG file for this set’s icon on
Scryfall’s CDN. Hotlinking this image isn’t recommended, because it may
change slightly over time. You should download it and use it locally for
your particular user interface needs.
}
}
\description{
Retrieve information about a set
}
\section{Set types}{

An exhaustive list of \code{set_types} is listed below:
\itemize{
\item \code{core}: A yearly Magic core set (Tenth Edition, etc)
\item \code{expansion}: A rotational expansion set in a block (Zendikar, etc)
\item \code{masters}: A reprint set that contains no new cards (Modern Masters, etc)
\item \code{masterpiece}: Masterpiece Series premium foil cards
\item \code{arsenal}: A Commander-oriented gift set
\item \code{from_the_vault}: From the Vault gift sets
\item \code{spellbook}: Spellbook series gift sets
\item \code{premium_deck}: Premium Deck Series decks
\item \code{duel_deck}: Duel Decks
\item \code{draft_innovation}: Special draft sets, like Conspiracy and Battlebond
\item \code{treasure_chest}: Magic Online treasure chest prize sets
\item \code{commander}: Commander preconstructed decks
\item \code{planechase}: Planechase sets
\item \code{archenemy}: Archenemy sets
\item \code{vanguard}: Vanguard card sets
\item \code{funny}: A funny un-set or set with funny promos (Unglued, Happy Holidays,
etc)
\item \code{starter}: A starter/introductory set (Portal, etc)
\item \code{box}: A gift box set
\item \code{promo}: A set that contains purely promotional cards
\item \code{token}: A set made up of tokens and emblems.
\item \code{memorabilia}: A set made up of gold-bordered, oversize, or trophy cards
that are not legal
}
}

