% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_cards.R
\name{scry_cards}
\alias{scry_cards}
\title{Query Scryfall Card Search API}
\usage{
scry_cards(query, .unique = "cards", .order = "name",
  direction = "auto", include_extras = FALSE,
  include_multilingual = FALSE, delay = 75)
}
\arguments{
\item{delay}{Number of microseconds scryr should wait between requests.
(Scryfall asks for 50-100)}
}
\value{
data frame of cards matching the search parameters
}
\description{
Function for contents of catalogs. Catalogs contains all
existing versions of that object. For example scry_catalog("toughnesses")
returns every toughness that has ever been printed on a \strong{Magic} card,
including non-mumeric values like "*+1". \cr
Available catalogs are "card-names", "artist-names", "word-bank",
"creature-types", "planeswalker-types", "land-types","artifact-types",
"enchantment-types", "spell-types", "powers", "toughnesses",
"loyalties", "watermarks". \cr
While most catalog names are self-explanatory, "word-bank" is every English
word of 2+ letters that's appears in a card name.
}
\examples{
scry_catalog("artist-names")
library(dplyr)
scry_catalog("toughnesses") \%>\% as.numeric() \%>\% unique()
}
