% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_catalog.R
\name{scry_sets}
\alias{scry_sets}
\title{Query Scryfall Catalog API}
\usage{
scry_sets(set_code = NULL, delay = 75)
}
\arguments{
\item{catalog_name}{name of catalog you want (character). See above for available catalogs.}
}
\value{
character vector of objects in the catalog
}
\description{
Function for contents of catalogs. Catalogs contains all
existing versions of that object. For example scry_catalog("toughnesses")
returns every toughness that has ever been printed on a \strong{Magic} card,
including non-mumeric values like "*+1".
Available catalogs are "card-names", "artist-names", "word-bank",
"creature-types", "planeswalker-types", "land-types","artifact-types",
"enchantment-types", "spell-types", "powers", "toughnesses",
"loyalties", "watermarks"
"word-bank" is every English word of 2+ letters that's appears in a card
name.
}
\examples{
scry_catalog("artist-names")
}
