% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_data.R
\name{bulk_rulings}
\alias{bulk_rulings}
\alias{bulk_default_cards}
\title{Fetch Scryfall Bulk Data}
\usage{
bulk_rulings()

bulk_default_cards()
}
\value{
\code{bulk_rulings} returns a \link[tibble:tibble-package]{tibble} of all
rulings as of the latest update. \code{oracle-id} column refers to the card
associate with the ruling
}
\description{
Scryfall releases bulk data files for cards and rulings that are
updated daily. These functions wrap calling those APIs. The cards bulk data
come in three versions: default (every card in English (unless only available
in a non-English language)), oracle (one card per Oracle ID), and all (every
card in every language). Documentation is available at
\url{https://scryfall.com/docs/api/bulk-data}.
}
\concept{bulk_data}
