% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_guild.R
\name{label_guild}
\alias{label_guild}
\title{Assign guild based on color/color identity}
\usage{
label_guild(color_code, inclusive = FALSE, convert_to_list = FALSE)
}
\arguments{
\item{color_code}{vector of characters of the card's color/color identity,
using WUBRG notation. Letters must be capital and in alphabetic order. If
using with \code{link{[dplyr]mutate}}, \code{link{[purrr]map_chr}}, and the
results of \code{link{scry_cards}}, use the \code{colors} or
\code{color_identity} column. See\code{vingette("using_label_functions")}.}

\item{inclusive}{if \code{color_code} is just one (or no) color, should
\code{label_guild} return all guilds that contain that color?}

\item{convert_to_list}{Should \code{label_guild} a list instead of a vector?
Useful if you're planning on using functions like \code{link[tidyr]unnest}.}
}
\value{
a vector of strings with all guild(s) matching the {color_code}
}
\description{
Given a card's color (or color identity), return the Ravinca
guild assoicated with the card. By default, \code{label_guild} requires the
card to be both of the guild's colors, but if \code{inclusive = TRUE},
mono-color cards are labeled with all guilds they could belong to.
}
\section{Note}{

cards with hybrid costs are treated as belonging to both colors, as though
they were gold cards.
}

\section{Warning}{

If you're using \code{label_guild} inside
\code{\link{[dplyr]mutate}}, you'll need to pair it with
\code{\link{[purrr]map}}. See\code{vingette("using_label_functions")}.
}

\examples{
label_guild(c("U", "W"))
label_guild("U")
label_guild("U", inclusive = TRUE)
}
\seealso{
\code{\link{label_tri}}
}
\concept{label}
