% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-cards.R
\name{scry-cards}
\alias{scry-cards}
\alias{scry_cards}
\title{Retrieve information about cards}
\usage{
scry_cards(
  q,
  unique = "cards",
  order = "name",
  dir = "auto",
  include_extras = FALSE,
  include_multilingual = FALSE,
  include_variations = FALSE,
  page = 1
)
}
\arguments{
\item{q}{A search query using the same \href{https://scryfall.com/docs/syntax}{fulltext search system} that the main site uses. Maximum
length: 1000 Unicode characters.}

\item{unique}{The strategy for omitting similar cards. See below.}

\item{order}{The method to sort returned cards. See below.}

\item{dir}{The direction to sort cards. See below.}

\item{include_extras}{If true, extra cards (tokens, planes, etc) will be
included. Equivalent to adding include:extras to the fulltext search.}

\item{include_multilingual}{If true, cards in every language supported by
Scryfall will be included.}

\item{include_variations}{If true, rare care variants will be included, like
the \href{https://scryfall.com/card/drk/107\%E2\%80\%A0/runesword}{Hairy Runesword}.}

\item{page}{The page number to return.}
}
\description{
Retrieve information about cards
}
\details{
Card objects represent individual Magic: The Gathering cards that
players could obtain and add to their collection (with a few minor
exceptions).

Cards are the API's most complex object. You are encouraged to thoroughly
read this document and also \code{vignette("layouts")}.
\subsection{Card Names}{

Internally, Scryfall tracks the uniqueness of "Oracle names." (i.e. names you
can pick when an effect asks you to "choose a card name"). Each unique Oracle
name is separately available in \code{\link[=catalog_card_names]{catalog_card_names()}}.

Note that while most Oracle card names are unique, Scryfall also indexes
other objects such as tokens and Unstable set variants which do not always
have a unique name.
}

\subsection{Multiface Cards}{

Magic cards can have multiple faces or multiple cards printed on one card
stock. The faces could be shown divided on the front of the card as in split
cards and flip cards, or the card can be double-faced as in transform cards
and modal DFCs.

Scryfall represents multi-face cards as a single object with a \code{card_faces}
array describing the distinct faces. More details below.
}
}
\section{Special Arguments}{

\subsection{Unique (Rollup) Modes}{

The \code{unique} parameter specifies if Scryfall should remove "duplicate"
results in your query. The options are:
\itemize{
\item \code{cards}: Removes duplicate gameplay objects (cards that share a name and
have the same functionality). For example, if your search matches more than
one print of Pacifism, only one copy of Pacifism will be returned.
\item \code{art}: Returns only one copy of each unique artwork for matching cards. For
example, if your search matches more than one print of Pacifism, one card
with each different illustration for Pacifism will be returned, but any
cards that duplicate artwork already in the results will be omitted.
\item \code{prints}: Returns all prints for all cards matched (disables rollup). For
example, if your search matches more than one print of Pacifism, all
matching prints will be returned.
}
}

\subsection{Sorting Cards}{

The \code{order} parameter determines how Scryfall should sort the returned cards.
\itemize{
\item \code{name}: Sort cards by name, A → Z.
\item \code{set}: Sort cards by their set and collector number: AAA/#1 → ZZZ/#999.
\item \code{released}: Sort cards by their release date: Newest → Oldest.
\item \code{rarity}: Sort cards by their rarity: Common → Mythic.
\item \code{color}: Sort cards by their color and color identity: WUBRG → multicolor →
colorless.
\item \code{usd}: "Sort cards by their lowest known U.S. Dollar price: 0.01 → highest,
null last".
\item \code{tix}: "Sort cards by their lowest known TIX price: 0.01 → highest, null
last".
\item \code{eur}: "Sort cards by their lowest known Euro price: 0.01 → highest, null
last".
\item \code{cmc}: Sort cards by their converted mana cost: 0 → highest.
\item \code{power}: Sort cards by their power: null → highest.
\item \code{toughness}: Sort cards by their toughness: null → highest.
\item \code{edhrec}: Sort cards by their EDHREC ranking: lowest → highest.
\item \code{artist}: Sort cards by their front-side artist name: A → Z.
\item \code{review}: Sort cards how podcasts review sets, usually color & CMC, lowest
→ highest, with Booster Fun cards at the end.
}

Then you can optionally specify a \code{dir} parameter to choose which direction
the sorting should occur:
\itemize{
\item \code{auto}: Scryfall will automatically choose the most inuitive direction to
sort.
\item \code{asc}: Sort ascending (the direction of the arrows in the previous table).
\item \code{desc}: Sort descending (flip the direction of the arrows in the previous
table).
}
}
}

\section{Special Columns}{

\subsection{Card Face Objects}{

Multiface cards have a \code{card_faces} property containing at least two Card
Face objects. Those objects have the following properties:
\itemize{
\item \code{object} [chr]: A content type for this object, always \code{card_face}.
\item \code{oracle_id} [chr]: The Oracle ID of this particular face, if the card is
reversible.
\item \code{cmc} [dbl]: The mana value of this particular face, if the card is
reversible.
\item \code{color_indicator} [lst]: The colors in this face's color indicator, if
any.
\item \code{colors} [lst]: This face's colors, if the game defines colors for the
individual face of this card.
\item \code{layout} [chr]: The layout of this card face, if the card is reversible.
\item \code{loyalty} [chr]: This face's loyalty, if any.
\item \code{mana_cost} [chr]: The mana cost for this face. This value will be any
empty string \code{""} if the cost is absent. Remember that per the game rules,
a missing mana cost and a mana cost of \code{{0}} are different values.
\item \code{name} [chr]: The name of this particular face.
\item \code{oracle_text} [chr]: The Oracle text for this face, if any.
\item \code{power} [chr]: This face's power, if any. Note that some cards have
powers that are not numeric, such as \code{*}.
\item \code{toughness} [chr]: This face's toughness, if any.
\item \code{type_line} [chr]: The type line of this particular face, if the card is
reversible.
\item \code{artist} [chr]: The name of the illustrator of this card face. Newly
spoiled cards may not have this field yet.
\item \code{artist_id} [chr]: The unique identifier of the illustrator of this card
face.
\item \code{flavor_name} [chr]: The just-for-fun name printed on this card face
(such as for Godzilla series cards).
\item \code{flavor_text} [chr]: The flavor text printed on this face, if any.
\item \code{illustration_id} [chr]: A unique identifier for the card face artwork
that remains consistent across reprints. Newly spoiled cards may not have
this field yet.
\item \code{image_uris} [tbl]: An object providing URIs to imagery for this face, if
this is a double-sided card. If this card is not double-sided, then the
\code{image_uris} property will be part of the parent object instead.
\item \code{printed_name} [chr]: The localized name printed on this face, if any.
\item \code{printed_text} [chr]: The localized text printed on this face, if any.
\item \code{printed_type_line} [chr]: The localized type line printed on this face,
if any.
\item \code{watermark} [chr]: The watermark on this particulary card face, if any.
}
}
}

