% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_cards.R
\name{scry_cards}
\alias{scry_cards}
\title{Query Scryfall Card Search API}
\usage{
scry_cards(query, .unique = "cards", .order = "name",
  direction = "auto", include_extras = FALSE,
  include_multilingual = FALSE, include_variations = FALSE,
  delay = 75)
}
\arguments{
\item{.unique}{How Scryfall handles cases where different versions of the
same card match the \code{query}. "cards" (default) returns only one instance of
card, "art" returns each instance with a different art "prints" returns all
of them.}

\item{.order}{How Scryfall sorts returned cards. "name" (default): card name,
"set": set code and collector number, "released": release date, "rarity":
rarity, "color": color, "usd": price in US dollars, "tix"; price in tickets
on MTGO, "eur": price in Euros, "cmc": converted mana cost, "power": power,
"toughenss": toughness, "edhrec": EDHREC rating, or "artist": artist name.
See vignette("sorting-cards") for more information on how sorting works.}

\item{direction}{Which direction cards are sorted in based on \code{.order}.
"auto" (default): order that makes most sense for \code{.order},
"asc": ascending order, and "desc": descending order. See
vignette("sorting-cards") for more information on how sorting works.}

\item{include_extras}{Should results include extras like tokens or schemes.
The default is FALSE.}

\item{include_multilingual}{Should results include cards in all supported
languages. The default is FALSE.}

\item{include_variations}{Should results include rare case variants. The
default is FALSE.}

\item{delay}{Number of milliseconds scryr should wait between requests.
(Scryfall asks for 50-100)}
}
\value{
a \link[tibble:tibble-package]{tibble} of cards matching the search
parameters
}
\description{
Function for contents of catalogs. Catalogs contains all
existing versions of that object. For example scry_catalog("toughnesses")
returns every toughness that has ever been printed on a \strong{Magic} card,
including non-mumeric values like "*+1". \cr
Available catalogs are "card-names", "artist-names", "word-bank",
"creature-types", "planeswalker-types", "land-types","artifact-types",
"enchantment-types", "spell-types", "powers", "toughnesses",
"loyalties", "watermarks". \cr
While most catalog names are self-explanatory, "word-bank" is every English
word of 2+ letters that's appears in a card name.
}
\examples{
scry_catalog("artist-names")
library(dplyr)
scry_catalog("toughnesses") \%>\% as.numeric() \%>\% unique()
}
