% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_cards.R
\name{scry_cards}
\alias{scry_cards}
\title{Query Scryfall Card Search API}
\usage{
scry_cards(query, .unique = "cards", .order = "name",
  direction = "auto", include_extras = FALSE,
  include_multilingual = FALSE, include_variations = FALSE,
  include_ids = FALSE, include_uris = FALSE, delay = 75)
}
\arguments{
\item{query}{search query to pass to the API. A character string that looks
the same as a search you'd manually enter into Scryfall.
(see: \url{https://scryfall.com/docs/syntax} for documentation on searching
Scryfall)}

\item{.unique}{How Scryfall handles cases where different versions of the
same card match the \code{query}. "cards" (default) returns only one instance of
card, "art" returns each instance with a different art "prints" returns all
of them.}

\item{.order}{How Scryfall sorts returned cards. "name" (default): card name,
"set": set code and collector number, "released": release date, "rarity":
rarity, "color": color, "usd": price in US dollars, "tix"; price in tickets
on MTGO, "eur": price in Euros, "cmc": converted mana cost, "power": power,
"toughenss": toughness, "edhrec": EDHREC rating, or "artist": artist name.}

\item{direction}{Which direction cards are sorted in based on \code{.order}.
"auto" (default): order that makes most sense for \code{.order},
"asc": ascending order, and "desc": descending order.  "auto" sorts in
descending order for prices ("usd", "tix", and "eur"), "release date", and
"rarity". For all other \code{.order} options, "auto sorts in ascending
order.}

\item{include_extras}{Should results include extras like tokens or schemes.}

\item{include_multilingual}{Should results include cards in all supported
languages.}

\item{include_variations}{Should results include rare case variants.}

\item{include_ids}{Should results include ID variables (e.g. Oracle ID)}

\item{include_uris}{Should results include URI variables (e.g. the URL for
the card on Scryfall)}

\item{delay}{Number of milliseconds scryr should wait between requests.
(Scryfall asks for 50-100)}
}
\value{
a \link[tibble:tibble-package]{tibble} of cards matching the search
parameters
}
\description{
Request card-level information. This API call operates in a
similar manner to a manual Scryfall search.
}
\examples{
scry_cards("set:ktk")
scry_cards("set:ktk wm:mardu")
}
\concept{search}
