% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_catalog.R
\name{scry_catalog}
\alias{scry_catalog}
\title{Query Scryfall Catalog API}
\usage{
scry_catalog(catalog_name, delay = 75)
}
\arguments{
\item{catalog_name}{The name of the catalog you want (character). See
Description for available catalogs.}

\item{delay}{Number of milliseconds scryr should wait between requests
(Scryfall asks for 50-100).}
}
\value{
character vector of objects in the catalog
}
\description{
Request contents of catalogs. Catalogs contains
all existing versions of that object. For example scry_catalog("toughnesses")
returns every toughness that has ever been printed on a \strong{Magic} card,
including non-mumeric values like "*+1".
}
\examples{
scry_catalog("artist-names")
scry_catalog("toughnesses") \%>\% as.numeric()
}
\seealso{
Catalogs:
Available catalogs are "card-names", "artist-names", "word-bank",
"creature-types", "planeswalker-types", "land-types","artifact-types",
"enchantment-types", "spell-types", "powers", "toughnesses",
"loyalties", "watermarks". \cr
While most catalog names are self-explanatory, "word-bank" is every English
word of 2+ letters that's appears in a card name.
}
\concept{search}
