% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_catalog.R
\name{scry_catalog}
\alias{scry_catalog}
\title{Query Scryfall Catalog API}
\usage{
scry_catalog(catalog_name, delay = 75)
}
\arguments{
\item{catalog_name}{name of catalog you want (character). See above for available catalogs.}

\item{delay}{how many microseconds should scryr wait between requests
(Scryfall asks for 50-100)}
}
\value{
character vector of objects in the catalog
}
\description{
Function for reuqesting contents of catalogs. Catalogs contains
all existing versions of that object. For example scry_catalog("toughnesses")
returns every toughness that has ever been printed on a \strong{Magic} card,
including non-mumeric values like "*+1". \cr
Available catalogs are "card-names", "artist-names", "word-bank",
"creature-types", "planeswalker-types", "land-types","artifact-types",
"enchantment-types", "spell-types", "powers", "toughnesses",
"loyalties", "watermarks". \cr
While most catalog names are self-explanatory, "word-bank" is every English
word of 2+ letters that's appears in a card name.
}
\examples{
scry_catalog("artist-names")
library(dplyr)
scry_catalog("toughnesses") \%>\% as.numeric() \%>\% unique()
}
