% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-sets.R
\name{scry_set}
\alias{scry_set}
\title{Returns a set with the given set code}
\usage{
scry_set(code)
}
\arguments{
\item{code}{The three to five-letter set code. Can be either the \code{code} or
the \code{mtgo_code} for the set.}
}
\description{
Returns a set with the given set code
}
