% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-sets.R
\name{scry_set}
\alias{scry_set}
\title{Returns a set with the given set code}
\usage{
scry_set(code)
}
\arguments{
\item{code}{The three to five-letter set code. Can be either the \code{code} or
the \code{mtgo_code} for the set.}
}
\value{
A data frame with 1 row and the following columns:
\itemize{
\item \code{id}: [UUID] A unique ID for this set on Scryfall that will not change.
\item \code{code}: [String] The unique three to five-letter code for this set.
\item \code{mtgo_code}: [String] The unique code for this set on MTGO, which may
differ from the regular code.
\item \code{tcgplayer_id}: [Integer] This set’s ID on TCGplayer’s API, also known as
the groupId.
\item \code{name}: [String] The English name of the set.
\item \code{set_type}: [String] A computer-readable classification for this set. See
below.
\item \code{released_at}: [Date] The date the set was released or the first card was
printed in the set (in GMT-8 Pacific time).
\item \code{block_code}: [String] Nullable  The block code for this set, if any.
\item \code{block}: [String] The block or group name code for this set, if any.
\item \code{parent_set_code}: [String] The set code for the parent set, if any.
promo and token sets often have a parent set.
\item \code{card_count}: [Integer] The number of cards in this set.
\item \code{printed_size}: [Integer] The denominator for the set’s printed collector
numbers.
\item \code{digital}: [Boolean] True if this set was only released in a video game.
\item \code{foil_only}: [Boolean] True if this set contains only foil cards.
\item \code{nonfoil_only}: [Boolean] True if this set contains only nonfoil cards.
\item \code{scryfall_uri}: [URI] A link to this set’s permapage on Scryfall’s
website.
\item \code{uri}: [URI] A link to this set object on Scryfall’s API.
\item \code{icon_svg_uri}: [URI] A URI to an SVG file for this set’s icon on
Scryfall’s CDN. Hotlinking this image isn’t recommended, because it may
change slightly over time. You should download it and use it locally for
your particular user interface needs.
\item \code{search_uri}: [URI] A Scryfall API URI that you can request to begin
paginating over the cards in this set.
}
}
\description{
Returns a set with the given set code
}
