% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry_sets.R
\name{scry_sets}
\alias{scry_sets}
\title{Query Scryfall Set API}
\usage{
scry_sets(set_code = NULL, delay = 75)
}
\arguments{
\item{set_code}{three letter set code. If NULL (default), returns all sets.}

\item{delay}{how many microseconds should scryr wait between requests
(Scryfall asks for 50-100)}
}
\value{
result of query (document of set object at https://scryfall.com/docs/api/sets)
}
\description{
Function for getting-set level information (e.g. release date,
number of cards). Formatted as a tibble.
}
\examples{
scry_sets() #returns all sets
scry_sets("aer") 
scry_sets("m19") # block-related columns absent
}
